requirements.txt
attrs==22.2.0
bokeh==2.4.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
dask==2023.3.2
distributed==2023.3.2
exceptiongroup==1.1.1
fsspec==2023.3.0
HeapDict==1.0.1
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
Jinja2==3.1.2
kaggle==1.5.13
locket==1.0.0
lz4==4.3.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.2
packaging==23.0
pandas==1.5.3
partd==1.3.0
Pillow==9.4.0
pluggy==1.0.0
psutil==5.9.4
pyarrow==11.0.0
pydantic==1.10.7
pytest==7.2.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.2
PyYAML==6.0
requests==2.28.2
six==1.16.0
sortedcontainers==2.4.0
tblib==1.7.0
text-unidecode==1.3
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
zict==2.2.0
zipp==3.15.0
